More tests (#330)
* Add some no-covers

* Refactor data tests

* Updates str and tests

* Add test

* Remove labels_from_categorical_csv

* Add test_base_viz

* Fix pep

* More tests

* Add tests for data pipeline state

* Remove old code

* Add test_properties.py

* Add tests for warnings and errors

* Move pragma

* B -> BATCH_SIZE
