--- flash/core/data/__init__.py
@@ -1,1 +1,0 @@
-from flash.core.data.data_utils import labels_from_categorical_csv
