+++ flash/core/data/auto_dataset.py
@@ -19,7 +19,7 @@ from torch.utils.data import Dataset, IterableDataset
 
 from flash.core.data.utils import CurrentRunningStageFuncContext
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no-cover
     from flash.core.data.data_pipeline import DataPipeline
     from flash.core.data.data_source import DataSource
