+++ tests/core/data/test_auto_dataset.py
