+++ tests/core/data/test_serialization.py
