+++ tests/core/data/test_split_dataset.py
