+++ tests/core/test_trainer.py
@@ -13,7 +13,10 @@
 # limitations under the License.
 from typing import Any
 
+import pytest
 import torch
+from pytorch_lightning.callbacks.early_stopping import EarlyStopping
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from torch import nn
 from torch.nn import functional as F
 
@@ -49,19 +52,61 @@ class DummyClassifier(nn.Module):
         return self.head(self.backbone(x))
 
 
-def test_task_fit(tmpdir: str):
+@pytest.mark.parametrize("callbacks, should_warn", [([], False), ([NoFreeze()], True)])
+def test_trainer_fit(tmpdir, callbacks, should_warn):
     model = nn.Sequential(nn.Flatten(), nn.Linear(28 * 28, 10), nn.LogSoftmax())
     train_dl = torch.utils.data.DataLoader(DummyDataset())
     val_dl = torch.utils.data.DataLoader(DummyDataset())
     task = ClassificationTask(model, loss_fn=F.nll_loss)
-    trainer = Trainer(fast_dev_run=True, default_root_dir=tmpdir)
-    trainer.fit(task, train_dl, val_dl)
+    trainer = Trainer(fast_dev_run=True, default_root_dir=tmpdir, callbacks=callbacks)
+
+    if should_warn:
+        with pytest.warns(UserWarning, match="trainer is using a fine-tuning callback"):
+            trainer.fit(task, train_dl, val_dl)
+    else:
+        trainer.fit(task, train_dl, val_dl)
 
 
-def test_task_finetune(tmpdir: str):
+def test_trainer_finetune(tmpdir):
     model = DummyClassifier()
     train_dl = torch.utils.data.DataLoader(DummyDataset())
     val_dl = torch.utils.data.DataLoader(DummyDataset())
     task = ClassificationTask(model, loss_fn=F.nll_loss)
     trainer = Trainer(fast_dev_run=True, default_root_dir=tmpdir)
     trainer.finetune(task, train_dl, val_dl, strategy=NoFreeze())
+
+
+def test_resolve_callbacks_invalid_strategy(tmpdir):
+    model = DummyClassifier()
+    trainer = Trainer(fast_dev_run=True, default_root_dir=tmpdir)
+    task = ClassificationTask(model, loss_fn=F.nll_loss)
+    with pytest.raises(MisconfigurationException, match="should be a ``pytorch_lightning.callbacks.BaseFinetuning``"):
+        trainer._resolve_callbacks(task, EarlyStopping())
+
+
+class MultiFinetuneClassificationTask(ClassificationTask):
+
+    def configure_finetune_callback(self):
+        return [NoFreeze(), NoFreeze()]
+
+
+def test_resolve_callbacks_multi_error(tmpdir):
+    model = DummyClassifier()
+    trainer = Trainer(fast_dev_run=True, default_root_dir=tmpdir)
+    task = MultiFinetuneClassificationTask(model, loss_fn=F.nll_loss)
+    with pytest.raises(MisconfigurationException, match="should create a list with only 1 callback"):
+        trainer._resolve_callbacks(task, None)
+
+
+class FinetuneClassificationTask(ClassificationTask):
+
+    def configure_finetune_callback(self):
+        return [NoFreeze()]
+
+
+def test_resolve_callbacks_override_warning(tmpdir):
+    model = DummyClassifier()
+    trainer = Trainer(fast_dev_run=True, default_root_dir=tmpdir)
+    task = FinetuneClassificationTask(model, loss_fn=F.nll_loss)
+    with pytest.warns(UserWarning, match="The model contains a default finetune callback"):
+        trainer._resolve_callbacks(task, "test")
